extract tracks.js, start producing FeatureCollection responses
