+++ README.md
@@ -17,15 +17,9 @@ If enabled by black/whitelist configuration `navigation.position` updates are wr
 
 The coordinates are written as `[lon, lat]` strings for minimal postprocessing in GeoJSON conversion.
 
-The plugin creates `/signalk/vX/api/self/track` endpoint that accepts two parameters:
-- timespan in the xxxY format, where xxx is a Number and Y one of 
-  - s  (seconds)
-  - m  (minutes)
-  - h  (hours)
-  - d  (days)
-  - w  (weeks)
-- resolution in the same format
-and returns GeoJSON MultiLineString. For example `http://localhost:3000/signalk/v1/api/self/track?timespan=1d&resolution=1h` will return the data for the last 1 day (24 hours) with one position per hour. The data is simply sampled with InfluxDB's `first()` function.
+The plugin creates `/signalk/vX/api/vessels/self/tracks` endpoint that accepts a single bbox parameter in `southwest_lng,southwest_lat,northeast_lng,northeast_lat` format and returns a GeoJSON FeatureCollection with one MultiLineString feature. The track feature has the following properties:
+- `name` (human oriented name)
+- `starttime` and `endtime` (ISO 8601 strings in Zulu time)
 
 ### Try it out
